Fix thinking messages
