+++ src/components/ChatWidget.tsx
@@ -16,26 +16,16 @@ import {
   Send as SendIcon,
   AttachFile as AttachFileIcon,
   Close as CloseIcon,
-  CheckCircle as CheckCircleIcon,
   Refresh as RefreshIcon,
-  Psychology as PsychologyIcon,
-  ExpandMore as ExpandMoreIcon,
 } from '@mui/icons-material';
-import { motion, AnimatePresence } from 'framer-motion';
-import ReactMarkdown from 'react-markdown';
+import { AnimatePresence } from 'framer-motion';
 import { useConversationStore } from '../store/conversationStore';
 import { apiClient } from '../api/client';
-import { config } from '../config';
 import { storage } from '../utils/storage';
+import { MessageBubble } from './MessageBubble';
 import type { Message } from '../api/types';
 
-interface ChatWidgetProps {
-  onConversationCreated: () => void;
-}
-
-const MotionBox = motion(Box);
-
-export const ChatWidget: React.FC<ChatWidgetProps> = ({ onConversationCreated }) => {
+export const ChatWidget: React.FC = () => {
   const {
     messages,
     models,
@@ -55,14 +45,72 @@ export const ChatWidget: React.FC<ChatWidgetProps> = ({ onConversationCreated })
   const [images, setImages] = useState<File[]>([]);
   const [isStreaming, setIsStreaming] = useState(false);
   const [streamingContent, setStreamingContent] = useState('');
+  const [streamingThinking, setStreamingThinking] = useState('');
   const [displayedContent, setDisplayedContent] = useState(''); // For typing effect
+  const [displayedThinking, setDisplayedThinking] = useState(''); // For typing effect
   const [justFinishedStreaming, setJustFinishedStreaming] = useState(false);
   const [expandedThinking, setExpandedThinking] = useState<Set<number>>(new Set()); // Track which messages have thinking expanded
+  const [activeConversationId, setActiveConversationId] = useState<number | null>(
+    currentConversation?.id || null
+  );
   const messagesEndRef = useRef<HTMLDivElement>(null);
   const messagesContainerRef = useRef<HTMLDivElement>(null);
   const fileInputRef = useRef<HTMLInputElement>(null);
   const previousScrollHeightRef = useRef<number>(0);
   const typingIntervalRef = useRef<NodeJS.Timeout | null>(null);
+  const streamFrameRef = useRef<number | null>(null);
+  const pendingStreamRef = useRef<{ content: string; thinking: string }>({ content: '', thinking: '' });
+
+  const cancelStreamUpdate = () => {
+    if (streamFrameRef.current !== null) {
+      cancelAnimationFrame(streamFrameRef.current);
+      streamFrameRef.current = null;
+    }
+  };
+
+  const scheduleStreamUpdate = (content: string, thinking: string) => {
+    pendingStreamRef.current = { content, thinking };
+    if (streamFrameRef.current === null) {
+      streamFrameRef.current = requestAnimationFrame(() => {
+        const next = pendingStreamRef.current;
+        setStreamingContent(next.content);
+        setStreamingThinking(next.thinking);
+        streamFrameRef.current = null;
+      });
+    }
+  };
+
+  // Sequential typing effect: thinking first, then content
+  const startSequentialTypingEffect = () => {
+    if (typingIntervalRef.current) {
+      clearInterval(typingIntervalRef.current);
+      typingIntervalRef.current = null;
+    }
+
+    typingIntervalRef.current = setInterval(() => {
+      setDisplayedThinking((prevThinking) => {
+        // First, type out thinking
+        if (prevThinking.length < streamingThinking.length) {
+          return streamingThinking.slice(0, prevThinking.length + 2);
+        }
+
+        // Thinking is done, now type content
+        setDisplayedContent((prevContent) => {
+          if (prevContent.length < streamingContent.length) {
+            return streamingContent.slice(0, prevContent.length + 2);
+          }
+          // Both done, clear interval
+          if (typingIntervalRef.current) {
+            clearInterval(typingIntervalRef.current);
+            typingIntervalRef.current = null;
+          }
+          return prevContent;
+        });
+
+        return prevThinking;
+      });
+    }, 20);
+  };
 
   useEffect(() => {
     loadModels().catch(console.error);
@@ -83,33 +131,27 @@ export const ChatWidget: React.FC<ChatWidgetProps> = ({ onConversationCreated })
     }
   }, [selectedModel]);
 
-  // Typing effect - display characters one by one
   useEffect(() => {
-    // Clear any existing interval
-    if (typingIntervalRef.current) {
-      clearInterval(typingIntervalRef.current);
-      typingIntervalRef.current = null;
+    setActiveConversationId(currentConversation?.id || null);
+  }, [currentConversation]);
+
+  // Sequential typing effect - thinking first, then content
+  useEffect(() => {
+    if (!isStreaming) {
+      // Not streaming, clear any typing effects
+      if (typingIntervalRef.current) {
+        clearInterval(typingIntervalRef.current);
+        typingIntervalRef.current = null;
+      }
+      return;
     }
 
-    // If streaming content is longer than displayed content, start typing
-    if (streamingContent.length > displayedContent.length) {
-      typingIntervalRef.current = setInterval(() => {
-        setDisplayedContent((prev) => {
-          if (prev.length < streamingContent.length) {
-            // Display 2 characters at a time for faster typing
-            return streamingContent.slice(0, prev.length + 2);
-          }
-          // Finished typing
-          if (typingIntervalRef.current) {
-            clearInterval(typingIntervalRef.current);
-            typingIntervalRef.current = null;
-          }
-          return prev;
-        });
-      }, 20); // 20ms interval for smooth typing effect
-    } else if (streamingContent === '' && displayedContent !== '') {
-      // Reset when streaming content is cleared
+    // Reset and start sequential typing when streaming data changes
+    if (streamingThinking === '' && streamingContent === '') {
+      setDisplayedThinking('');
       setDisplayedContent('');
+    } else {
+      startSequentialTypingEffect();
     }
 
     return () => {
@@ -118,7 +160,13 @@ export const ChatWidget: React.FC<ChatWidgetProps> = ({ onConversationCreated })
         typingIntervalRef.current = null;
       }
     };
-  }, [streamingContent]);
+  }, [streamingThinking, streamingContent, isStreaming]);
+
+  useEffect(() => {
+    return () => {
+      cancelStreamUpdate();
+    };
+  }, []);
 
   // Reset "just finished" indicator after 3 seconds
   useEffect(() => {
@@ -133,9 +181,9 @@ export const ChatWidget: React.FC<ChatWidgetProps> = ({ onConversationCreated })
   // Scroll to bottom on new messages (but not when loading more old messages)
   useEffect(() => {
     if (!isLoadingMessages) {
-      messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
+      messagesEndRef.current?.scrollIntoView({ behavior: isStreaming ? 'auto' : 'smooth' });
     }
-  }, [messages, displayedContent, isLoadingMessages]);
+  }, [messages, displayedContent, isLoadingMessages, isStreaming]);
 
   // Preserve scroll position after loading more messages
   useEffect(() => {
@@ -183,59 +231,51 @@ export const ChatWidget: React.FC<ChatWidgetProps> = ({ onConversationCreated })
       };
 
       addMessage(userMessage);
+      addMessage({ role: 'assistant', content: '' });
       setInput('');
       setImages([]);
       setStreamingContent('');
+      setStreamingThinking('');
       setDisplayedContent(''); // Reset typing effect
+      setDisplayedThinking('');
       setJustFinishedStreaming(false); // Clear previous "done" indicator
 
       // Will start with empty message from backend
       let currentRole: string | null = null;
+      let hasPlaceholder = true;
       let accumulatedContent = '';
       let accumulatedThinking = ''; // Track thinking
-
-      console.log('[ChatWidget] Starting stream...');
 
       const stream = apiClient.chatStream(
         userMessage.content,
         selectedModel,
-        currentConversation?.id || null,
+        activeConversationId,
         imageFiles
       );
 
-      for await (const chunk of stream) {
-        console.log('[ChatWidget] Received chunk:', chunk);
+      let streamConversationId = activeConversationId;
 
-        // Handle conversation creation
-        if (chunk.conversation_id && !currentConversation) {
-          console.log('[ChatWidget] Setting conversation ID:', chunk.conversation_id);
-          setCurrentConversationId(chunk.conversation_id);
-          onConversationCreated();
+      for await (const chunk of stream) {
+        if (chunk.conversation_id && !streamConversationId) {
+          streamConversationId = chunk.conversation_id;
+          setActiveConversationId(chunk.conversation_id);
+          // Notify store to refresh conversation list if this is a new conversation
+          setCurrentConversationId(chunk.conversation_id).catch(console.error);
         }
 
         // Handle "done" signal from backend
         if (chunk.done) {
-          console.log('[ChatWidget] Received DONE signal from backend');
           // When streaming ends, update the store with final content and thinking
           if (accumulatedContent || accumulatedThinking) {
-            const finalMessage: Partial<Message> = { content: accumulatedContent };
-            if (accumulatedThinking) {
-              finalMessage.thinking = accumulatedThinking;
-            }
-            updateLastMessage(accumulatedContent);
-            // Update the last message with thinking if present
-            if (accumulatedThinking && messages.length > 0) {
-              const lastMessage = messages[messages.length - 1];
-              lastMessage.thinking = accumulatedThinking;
-            }
-            console.log('[ChatWidget] Updated last message in store with final content and thinking');
+            updateLastMessage(accumulatedContent, accumulatedThinking || undefined);
           }
 
           // Brief delay before showing "done" indicator and clearing streaming state
           setTimeout(() => {
+            cancelStreamUpdate();
             setStreamingContent('');
+            setStreamingThinking('');
             setJustFinishedStreaming(true);
-            console.log('[ChatWidget] Cleared streaming state, showing done indicator');
           }, 300);
           break; // Exit the loop when done
         }
@@ -243,15 +283,11 @@ export const ChatWidget: React.FC<ChatWidgetProps> = ({ onConversationCreated })
         // Handle message chunks
         if (chunk.message) {
           const messageRole = chunk.message.role;
-          console.log('[ChatWidget] Message role:', messageRole, 'Content:', chunk.message.content);
 
           // If role changes (e.g., assistant -> tool, or assistant -> agent), create new message
           if (currentRole && messageRole !== currentRole) {
-            console.log('[ChatWidget] Role changed from', currentRole, 'to', messageRole);
             // Finalize previous message
-            if (accumulatedContent) {
-              updateLastMessage(accumulatedContent);
-            }
+            updateLastMessage(accumulatedContent, accumulatedThinking || undefined);
 
             // Start new message for different role (empty content initially)
             const newMessage: Message = {
@@ -261,49 +297,55 @@ export const ChatWidget: React.FC<ChatWidgetProps> = ({ onConversationCreated })
             addMessage(newMessage);
             currentRole = messageRole;
             accumulatedContent = chunk.message.content || '';
+            accumulatedThinking = '';
+            setDisplayedThinking('');
 
             // Update streaming content immediately
-            setStreamingContent(accumulatedContent);
-            console.log('[ChatWidget] Set streaming content (new role):', accumulatedContent);
+            scheduleStreamUpdate(accumulatedContent, accumulatedThinking);
           } else if (!currentRole) {
-            console.log('[ChatWidget] First message chunk, role:', messageRole);
-            // First message chunk - create initial message (empty content initially)
-            const firstMessage: Message = {
-              role: messageRole,
-              content: '' // Start empty - will display from streaming state
-            };
-            addMessage(firstMessage);
+            // First message chunk - reuse placeholder bubble
             currentRole = messageRole;
             accumulatedContent = chunk.message.content || '';
+            accumulatedThinking = '';
+            setDisplayedThinking('');
+
+            if (hasPlaceholder) {
+              updateLastMessage(accumulatedContent, accumulatedThinking || undefined, messageRole);
+              hasPlaceholder = false;
+            } else {
+              const firstMessage: Message = {
+                role: messageRole,
+                content: '' // Start empty - will display from streaming state
+              };
+              addMessage(firstMessage);
+            }
 
             // Update streaming content immediately
-            setStreamingContent(accumulatedContent);
-            console.log('[ChatWidget] Set streaming content (first):', accumulatedContent);
+            scheduleStreamUpdate(accumulatedContent, accumulatedThinking);
           } else {
             // Same role, accumulate content
             if (chunk.message.content) {
               accumulatedContent += chunk.message.content;
-              setStreamingContent(accumulatedContent);
-              console.log('[ChatWidget] Accumulated content length:', accumulatedContent.length);
+              scheduleStreamUpdate(accumulatedContent, accumulatedThinking);
             }
           }
 
           // Always accumulate thinking regardless of role logic (simple append to thinking section)
           if (chunk.message.thinking) {
             accumulatedThinking += chunk.message.thinking;
-            console.log('[ChatWidget] Accumulated thinking, length:', accumulatedThinking.length);
+            scheduleStreamUpdate(accumulatedContent, accumulatedThinking);
           }
         }
       }
-
-      console.log('[ChatWidget] Stream ended, final content length:', accumulatedContent.length);
     } catch (err: any) {
       console.error('Chat error:', err);
       // Only update if we have a message to update
       if (messages.length > 0) {
         updateLastMessage('Error: ' + (err.message || 'Failed to send message'));
       }
+      cancelStreamUpdate();
       setStreamingContent('');
+      setStreamingThinking('');
     } finally {
       setIsStreaming(false);
     }
@@ -396,283 +438,22 @@ export const ChatWidget: React.FC<ChatWidgetProps> = ({ onConversationCreated })
         )}
 
         <AnimatePresence>
-          {messages.map((message, index) => {
-            const isLast = index === messages.length - 1;
-            const isStreamingThisMessage = isLast && message.role !== 'user' && isStreaming;
-            const showFinishedIndicator = isLast && message.role !== 'user' && justFinishedStreaming && !isStreaming;
-
-            // IMPORTANT: During streaming, display from typing effect (displayedContent), NOT from database (message.content)
-            // This ensures we show live-generated content with typing animation
-            const displayContent = isStreamingThisMessage ? displayedContent : message.content;
-
-            console.log('[ChatWidget] Rendering message', index, 'isStreamingThisMessage:', isStreamingThisMessage, 'displayContent length:', displayContent?.length);
-
-            // Determine message styling based on role
-            const isUser = message.role === 'user';
-
-            // Use role name for label (capitalize first letter)
-            const label = isUser
-              ? 'You'
-              : message.role.charAt(0).toUpperCase() + message.role.slice(1);
-
-            // Color scheme for different roles
-            const getColorScheme = () => {
-              if (isUser) {
-                return {
-                  bg: '#FFFFFF',
-                  border: '#E5E5E5',
-                  label: 'text.primary'
-                };
-              }
-              // For all non-user roles (assistant, tool, agents, etc.)
-              switch (message.role) {
-                case 'tool':
-                  return {
-                    bg: '#FFF8E1',
-                    border: '#FFB74D',
-                    label: '#F57C00'
-                  };
-                case 'assistant':
-                default:
-                  return {
-                    bg: '#F0FDF9',
-                    border: '#10A37F33',
-                    label: 'primary.main'
-                  };
-              }
-            };
-
-            const colorScheme = getColorScheme();
-
-            return (
-              <MotionBox
-                key={index}
-                initial={{ opacity: 0, y: 20 }}
-                animate={{ opacity: 1, y: 0 }}
-                exit={{ opacity: 0 }}
-                transition={{ duration: 0.3 }}
-                sx={{
-                  mb: 2,
-                  display: 'flex',
-                  justifyContent: isUser ? 'flex-end' : 'flex-start',
-                }}
-              >
-                <Paper
-                  elevation={0}
-                  sx={{
-                    maxWidth: '80%',
-                    p: 2,
-                    backgroundColor: colorScheme.bg,
-                    border: '1px solid',
-                    borderColor: colorScheme.border,
-                  }}
-                >
-                  <Typography
-                    variant="caption"
-                    sx={{
-                      fontWeight: 600,
-                      color: colorScheme.label,
-                      mb: 1,
-                      display: 'block',
-                    }}
-                  >
-                    {label}
-                  </Typography>
-                  <Box
-                    sx={{
-                      '& p': { margin: 0, marginBottom: 1 },
-                      '& p:last-child': { marginBottom: 0 },
-                      '& code': {
-                        backgroundColor: '#F3F4F6',
-                        padding: '2px 6px',
-                        borderRadius: 1,
-                        fontFamily: 'Consolas, Monaco, monospace',
-                        fontSize: '0.875em',
-                      },
-                      '& pre': {
-                        backgroundColor: '#F3F4F6',
-                        padding: 2,
-                        borderRadius: 1,
-                        overflow: 'auto',
-                      },
-                    }}
-                  >
-                    {/* Thinking Section - Show FIRST */}
-                    {(message.thinking || isStreamingThisMessage) && (
-                      <Box sx={{ mb: displayContent || message.images ? 1 : 0, pb: displayContent || message.images ? 1 : 0, borderBottom: displayContent || message.images ? '1px solid' : 'none', borderColor: 'divider' }}>
-                        <Button
-                          size="small"
-                          startIcon={<PsychologyIcon />}
-                          endIcon={
-                            <ExpandMoreIcon
-                              sx={{
-                                transform: expandedThinking.has(index) ? 'rotate(180deg)' : 'rotate(0deg)',
-                                transition: 'transform 0.2s',
-                              }}
-                            />
-                          }
-                          onClick={() => toggleThinking(index)}
-                          sx={{
-                            textTransform: 'none',
-                            color: 'text.secondary',
-                            fontSize: '0.75rem',
-                            minWidth: 'auto',
-                            px: 1,
-                            py: 0.5,
-                          }}
-                        >
-                          Thinking
-                        </Button>
-                        {expandedThinking.has(index) && (
-                          <Box
-                            component={motion.div}
-                            initial={{ opacity: 0, height: 0 }}
-                            animate={{ opacity: 1, height: 'auto' }}
-                            exit={{ opacity: 0, height: 0 }}
-                            transition={{ duration: 0.2 }}
-                            sx={{
-                              mt: 1,
-                              p: 1.5,
-                              backgroundColor: 'rgba(0, 0, 0, 0.02)',
-                              borderRadius: 1,
-                              fontSize: '0.875rem',
-                              color: 'text.secondary',
-                              fontFamily: 'monospace',
-                              whiteSpace: 'pre-wrap',
-                              wordBreak: 'break-word',
-                              maxHeight: '400px',
-                              overflow: 'auto',
-                            }}
-                          >
-                            {message.thinking}
-                          </Box>
-                        )}
-                      </Box>
-                    )}
-                    {/* Show typing indicator if streaming but no content yet */}
-                    {isStreamingThisMessage && !displayContent && (
-                      <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                        <Typography variant="body2" color="text.secondary">
-                          {label} is typing
-                        </Typography>
-                        <Box sx={{ display: 'flex', gap: 0.5 }}>
-                          {[0, 1, 2].map((i) => (
-                            <Box
-                              key={i}
-                              sx={{
-                                width: 6,
-                                height: 6,
-                                borderRadius: '50%',
-                                backgroundColor: 'primary.main',
-                                animation: 'bounce 1.4s infinite ease-in-out',
-                                animationDelay: `${i * 0.16}s`,
-                                '@keyframes bounce': {
-                                  '0%, 80%, 100%': {
-                                    transform: 'scale(0)',
-                                    opacity: 0.5,
-                                  },
-                                  '40%': {
-                                    transform: 'scale(1)',
-                                    opacity: 1,
-                                  },
-                                },
-                              }}
-                            />
-                          ))}
-                        </Box>
-                      </Box>
-                    )}
-                    {/* Show images if present */}
-                    {message.images && message.images.length > 0 && (
-                      <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 1, mb: displayContent ? 1 : 0 }}>
-                        {message.images.map((imageUuid, idx) => (
-                          <Box
-                            key={idx}
-                            component="img"
-                            src={apiClient.getImageUrl(imageUuid)}
-                            alt={`Image ${idx + 1}`}
-                            sx={{
-                              maxWidth: '100%',
-                              maxHeight: 300,
-                              borderRadius: 1,
-                              cursor: 'pointer',
-                            }}
-                            onClick={() => window.open(apiClient.getImageUrl(imageUuid), '_blank')}
-                          />
-                        ))}
-                      </Box>
-                    )}
-                    {/* Show content with typing effect */}
-                    {displayContent && (
-                      <>
-                        <ReactMarkdown
-                          components={{
-                            img: ({ node, ...props }) => (
-                              <Box
-                                component="img"
-                                {...props}
-                                sx={{
-                                  maxWidth: '100%',
-                                  maxHeight: 400,
-                                  borderRadius: 1,
-                                  cursor: 'pointer',
-                                  my: 1,
-                                }}
-                                onClick={() => window.open(props.src, '_blank')}
-                              />
-                            ),
-                          }}
-                        >
-                          {displayContent}
-                        </ReactMarkdown>
-                        {isStreamingThisMessage && (
-                          <Box
-                            component="span"
-                            sx={{
-                              display: 'inline-block',
-                              width: '8px',
-                              height: '16px',
-                              backgroundColor: 'primary.main',
-                              marginLeft: '2px',
-                              animation: 'blink 1s infinite',
-                              '@keyframes blink': {
-                                '0%, 49%': { opacity: 1 },
-                                '50%, 100%': { opacity: 0 },
-                              },
-                            }}
-                          />
-                        )}
-                      </>
-                    )}
-                    {showFinishedIndicator && (
-                      <Box
-                        component={motion.div}
-                        initial={{ opacity: 0, scale: 0.8 }}
-                        animate={{ opacity: 1, scale: 1 }}
-                        exit={{ opacity: 0 }}
-                        transition={{ duration: 0.3 }}
-                        sx={{
-                          display: 'flex',
-                          alignItems: 'center',
-                          gap: 0.5,
-                          mt: 1,
-                          pt: 1,
-                          borderTop: '1px solid',
-                          borderColor: 'divider',
-                          color: 'success.main',
-                        }}
-                      >
-                        <CheckCircleIcon sx={{ fontSize: 16 }} />
-                        <Typography variant="caption" color="success.main">
-                          Done
-                        </Typography>
-                      </Box>
-                    )}
-                  </Box>
-                </Paper>
-              </MotionBox>
-            );
-          })}
+          {messages.map((message, index) => (
+            <MessageBubble
+              key={index}
+              message={message}
+              index={index}
+              isLast={index === messages.length - 1}
+              isStreaming={isStreaming}
+              streamingThinking={streamingThinking}
+              streamingContent={streamingContent}
+              displayedThinking={displayedThinking}
+              displayedContent={displayedContent}
+              justFinishedStreaming={justFinishedStreaming}
+              expandedThinking={expandedThinking}
+              onToggleThinking={toggleThinking}
+            />
+          ))}
         </AnimatePresence>
         <div ref={messagesEndRef} />
       </Box>
